requirements.txt
openai-agents
openai-agents[voice]
openai
python-dotenv
textual
textual-dev
numpy
sounddevice
PyYAML